Refactor - move PostComment into article folder and change related imports
